move identical functions to hcshared.py and make ctx is print is indented correctly
